Fixed compilation issue on 64 bit platform.
git-svn-id: https://aug.svn.sourceforge.net/svnroot/aug/trunk@87 a3a97f7c-8e16-0410-9893-1d5f7d0f8f6e

+++ src/c/augnet/http.h
@@ -13,8 +13,8 @@ struct aug_var;
 struct aug_httphandlers {
     void (*setinitial_)(const struct aug_var*, const char*);
     void (*setfield_)(const struct aug_var*, const char*, const char*);
-    void (*setcsize_)(const struct aug_var*, size_t);
-    void (*cdata_)(const struct aug_var*, const void*, size_t);
+    void (*setcsize_)(const struct aug_var*, unsigned);
+    void (*cdata_)(const struct aug_var*, const void*, unsigned);
     void (*end_)(const struct aug_var*, int);
 };
